# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.User:
    properties:
      id_user:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  entity.UserErrorResponse:
    properties:
      error:
        example: Invalid product
        type: string
    type: object
  entity.UserReqUpdate:
    properties:
      name:
        type: string
    type: object
  entity.UserResponse:
    properties:
      id_user:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Pulsa Application
  title: Server Pulsa API
  version: "1.0"
paths:
...
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user by its ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/entity.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.UserReqUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user
          schema:
            $ref: '#/definitions/entity.UserResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users:
    get:
      consumes:
      - application/json
      description: Get a list of all users
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              items:
                $ref: '#/definitions/entity.User'
              type: array
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.UserErrorResponse'
      security:
      - BearerAuth: []
      summary: List all Users
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"